refactor: 상태관리 로직을 훅으로 분리 #16
