+++ app/_components/TopicSelector.tsx
@@ -1,59 +1,26 @@
 'use client';
 
 import clsx from 'clsx';
-import { useRouter } from 'next/navigation';
-import { useState } from 'react';
 
+import useTopicSelector from '@/app/_hooks/useTopicSelector';
 import { DeveloperType } from '@/app/_types/interview';
 
 import SelectButton from './SelectButton';
 import styles from './TopicSelector.module.scss';
 
-interface TopicSelectorProps {
+export interface TopicSelectorProps {
   variant?: 'topic' | 'subTopic';
   devType: DeveloperType;
   topics: string[];
 }
 
 export default function TopicSelector({ variant = 'topic', devType, topics }: TopicSelectorProps) {
-  const router = useRouter();
-  const [selectedTopics, setSelectedTopics] = useState<string[]>([]);
-  const isAllSelected = selectedTopics.length === topics.length;
-
-  const navigateToNext = () => {
-    if (!selectedTopics.length) return;
-
-    switch (variant) {
-      case 'topic':
-        const topicParam = isAllSelected ? 'all' : selectedTopics.join(',');
-        router.push(`/interview/select/${devType}/prepare?topics=${topicParam}`);
-        break;
-      case 'subTopic':
-        router.push('/interview/chat');
-        break;
-    }
-  };
-
-  const toggleTopic = (prevTopics: string[], clickedTopic: string) => {
-    if (prevTopics.includes(clickedTopic)) {
-      const newSelectedTopics = prevTopics.filter((topic) => topic !== clickedTopic);
-      return newSelectedTopics;
-    }
-    return [...prevTopics, clickedTopic];
-  };
-
-  const handleClickTopic = (clickedTopic: string) => {
-    setSelectedTopics((prevTopics) => toggleTopic(prevTopics, clickedTopic));
-  };
-
-  const selectAll = () => {
-    setSelectedTopics((prevSelectedTopics) => {
-      if (prevSelectedTopics.length === topics.length) return [];
-      return topics;
+  const { selectedTopics, isAllSelected, handleClickTopic, selectAll, isSelectedTopic, navigateToNext } =
+    useTopicSelector({
+      variant,
+      devType,
+      topics,
     });
-  };
-
-  const isSelectedTopic = (topic: string) => selectedTopics.includes(topic);
 
   return (
     <>
